pointer protection and optimisation

+++ srcs/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: lowatell <lowatell@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/06 03:03:54 by lowatell          #+#    #+#             */
-/*   Updated: 2024/10/08 10:51:13 by lowatell         ###   ########.fr       */
+/*   Updated: 2024/10/15 22:47:33 by lowatell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
-	while (i < (int)n)
+	while (i < n)
 	{
 		if ((unsigned char)c == ((unsigned char *)s)[i])
 			return ((unsigned char *)s + i);
