+++ srcs/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: lowatell <lowatell@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/28 06:01:42 by lowatell          #+#    #+#             */
-/*   Updated: 2024/10/08 10:51:17 by lowatell         ###   ########.fr       */
+/*   Updated: 2024/10/15 22:47:17 by lowatell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	if (!dst && !src)
 		return (NULL);
 	i = 0;
-	while (i < (int)n)
+	while (i < n)
 	{
 		((unsigned char *)dst)[i] = ((unsigned char *)src)[i];
 		i++;
